Move previous hooks implementation to non persistent core 2.0 hooks Everything is practically in place apart from actual hook behaviour for Topic Provider (Comments) while other like bbcode bbfilter should be in working condition.
