+++ Hooks/PostSubBundle.php
@@ -12,6 +12,9 @@
 
 namespace Zikula\DizkusModule\Hooks;
 
+use Zikula\Common\Translator\TranslatorInterface;
+use Zikula\Bundle\HookBundle\Category\UiHooksCategory;
+
 /**
  * PostSubBundle
  *
@@ -19,20 +22,53 @@ namespace Zikula\DizkusModule\Hooks;
  */
 class PostSubBundle extends AbstractSubBundle
 {
-    public function __construct($title = '')
+    const EDIT_DISPLAY = 'dizkus.ui_hooks.post.display_view';
+    const EDIT_FORM = 'dizkus.ui_hooks.post.form_edit';
+    const EDIT_VALIDATE = 'dizkus.ui_hooks.post.validate_edit';
+    const EDIT_PROCESS = 'dizkus.ui_hooks.post.process_edit';
+    const DELETE_FORM = 'dizkus.ui_hooks.post.form_delete';
+    const DELETE_VALIDATE = 'dizkus.ui_hooks.post.validate_delete';
+    const DELETE_PROCESS = 'dizkus.ui_hooks.post.process_delete';
+
+    /**
+     * @var TranslatorInterface
+     */
+    private $translator;
+
+    /**
+     * @param TranslatorInterface $translator
+     */
+    public function __construct(TranslatorInterface $translator)
     {
-        $owner = 'ZikulaDizkusModule';
-        $area = 'subscriber.dizkus.ui_hooks.post';
-        $category = 'ui_hooks';
+        $this->translator = $translator;
+        parent::__construct();
+    }
 
-        parent::__construct($owner, $area, $category, $title);
+    public function getOwner()
+    {
+        return 'ZikulaDizkusModule';
+    }
 
-        $this->addEvent('display_view', 'dizkus.ui_hooks.post.ui_view');
-        $this->addEvent('form_edit', 'dizkus.ui_hooks.post.ui_edit');
-        $this->addEvent('form_delete', 'dizkus.ui_hooks.post.ui_delete');
-        $this->addEvent('validate_edit', 'dizkus.ui_hooks.post.validate_edit');
-        $this->addEvent('validate_delete', 'dizkus.ui_hooks.post.validate_delete');
-        $this->addEvent('process_edit', 'dizkus.ui_hooks.post.process_edit');
-        $this->addEvent('process_delete', 'dizkus.ui_hooks.post.process_delete');
+    public function getCategory()
+    {
+        return UiHooksCategory::NAME;
+    }
+
+    public function getTitle()
+    {
+        return $this->translator->__('Dizkus post subscriber');
+    }
+
+    public function getEvents()
+    {
+        return [
+            UiHooksCategory::TYPE_DISPLAY_VIEW => self::EDIT_DISPLAY,
+            UiHooksCategory::TYPE_FORM_EDIT => self::EDIT_FORM,
+            UiHooksCategory::TYPE_VALIDATE_EDIT => self::EDIT_VALIDATE,
+            UiHooksCategory::TYPE_PROCESS_EDIT => self::EDIT_PROCESS,
+            UiHooksCategory::TYPE_FORM_DELETE => self::DELETE_FORM,
+            UiHooksCategory::TYPE_VALIDATE_DELETE => self::DELETE_VALIDATE,
+            UiHooksCategory::TYPE_PROCESS_DELETE => self::DELETE_PROCESS,
+        ];
     }
 }
