+++ Resources/views/Form/hooks.widget.html.twig
@@ -94,7 +94,7 @@
     {% spaceless %}
         {% set binding_id = 'area' ~ random() %}
         <div class="list-group-item {% if value.enabled %}list-group-item-success{% else %}list-group-item-default{% endif %} clearfix">
-            <span class="text-info small">{{ __('Area:') }} <strong>{{ value.subscriber.area }}</strong></span>
+            <span class="text-info small">{{ __('Area:') }} <strong>{{ value.subscriber }}</strong></span>
             <div class="pull-right text-right">
                 {% if form.children.enabled is defined %}
                     {{ form_widget(form.children.enabled, {'attr': {'class': 'btn btn-primary radio-inline btn-xs' }}) }}
@@ -207,7 +207,7 @@
     {% spaceless %}
         {% set binding_id = 'area' ~ random() %}
         <div class="list-group-item {% if value.enabled %}list-group-item-success{% else %}list-group-item-default{% endif %} clearfix">
-            <span class="text-info small">{{ __('Area:') }} <strong>{{ value.provider.area }}</strong></span>
+            <span class="text-info small">{{ __('Area:') }} <strong>{{ value.provider }}</strong></span>
             <div class="pull-right text-right">
                 {% if form.children.enabled is defined %}
                     {{ form_widget(form.children.enabled, {'attr': {'class': 'btn btn-primary radio-inline btn-xs' }}) }}
